build: webpack

+++ src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   const videos = await Video.find({}).sort({ createdAt: "desc" });
@@ -7,31 +8,44 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+
   if (!video) {
     return res.status(404).render("404", { pageTitle: "Video not found" });
   }
+  const { owner } = video;
+
   return res.render("watch", {
     pageTitle: video.title,
     video,
+    owner,
   });
 };
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     console.log(video);
     return res.render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
   return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
-  const isVideoExists = await Video.exists({ _id: id });
-  if (!isVideoExists) {
+  const video = await Video.exists({ _id: id });
+  if (!video) {
     return res.redirect("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
   }
 
   const { title, description, hashtags } = req.body;
@@ -48,17 +62,26 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+
+  const { path: fileUrl } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    const video = new Video({
+    const video = await Video.create({
       title,
       description,
+      fileUrl,
+      owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
-    await video.save();
+    const user = await User.findById(_id);
+    user.videos.push(video._id);
+    user.save();
     console.log(`Upload video ${video}`);
   } catch (error) {
-    console.error(error._message);
+    console.log(error._message);
     return res.render("upload", {
       pageTitle: "Upload Video",
       errorMessage: error._message,
@@ -70,8 +93,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
-
   return res.redirect("/");
 };
